requirements.txt
streamlit
pandas
pillow
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
PyYAML>=6.0
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
